Refactor Lightning components structure
