+++ packages/example/src/components/Lightning/ImportWallet/model.ts
@@ -119,6 +119,7 @@ export class LightningImportWalletModel {
     if(!this.loginAndPassword){
       return
     }
+    this.setImportResult(ImportWalletResult.Init);
 
     try {
       this.setIsImporting(true);
