+++ packages/example/src/components/Lightning/UserGuide/index.ts
@@ -1,0 +1,1 @@
+export { Tooltip as UserGuide } from './Tooltip';
