+++ packages/example/src/components/Lightning/index.tsx
@@ -1,19 +1,3 @@
-import { observer } from 'mobx-react-lite';
-import CreateWallet from './CreateWallet';
-import Ready from './Ready';
-import { useAppStore } from '../../mobx';
-import { LNWalletStepStatus } from '../../mobx/user';
-
-const SetupLightning = observer(() => {
-  const {
-    user: { LNWalletStep },
-  } = useAppStore();
-
-  return (
-    <>
-      {LNWalletStep === LNWalletStepStatus.Ready && <Ready />}
-    </>
-  );
-});
-
-export default SetupLightning;
+export { Ready as LNSetupModal } from './Ready';
+export { PayInvoice } from './PayInvoice';
+export { UserGuide } from './UserGuide';
